Changed ProjectInfo.dissipative from _Dissipative class to dictionary and deleted the _Dissipative class
This was a TODO in the project_info.py code. Since it is now a dictionary and not a class, instead of using for example
dissipative.dielectrics_bulk = ... You now need to write  dissipative['dielectrics_bulk'] = ... I used a command to replace all
of the instances in the project with the new format, but unfortunatly, this could break the code for some people so I need to add some code
that makes it so if someone used this code like in the old way it would still work, but give them `deprecated` message. Until then this branch
should not be merged.

I wrote this code from my own laptop where I dont have access to HFSS software so I have no way of testing it right now so it might not work
at all. Tommorow I'll check it when I get to the lab