pinfo: Added warnings to _Dissipative and fixed the code
Using pinfo.dissipative is now much cleaner to the end used (it could be easilly represented in a script or a ipython notebook). Also added error messages if the user does something wrong. Fixed a small bug that happened when saving. This code is now stable and I ran tests on it to check it. I will add a unit test soon